2D 엔진 렌더 및 기본기능 탑제 완료

+++ GameSource/Opengl2Md2/IGameStage.h
@@ -7,7 +7,7 @@ public:
 	virtual void onDrawScreen() = 0;
 
 	virtual void onTouchEvent(MotionEvent event) = 0;
-	virtual void onUpdate(long dt) = 0;
+	virtual void onUpdate(long long dt) = 0;
 	virtual void onClose() = 0;
 	virtual bool isCloseed() = 0;
 };
